applicant organization ownership checker added

+++ server_api/src/modules/applicants/applicant.service.ts
@@ -4,6 +4,7 @@ import { Gender, Accessibility, Scope } from "./applicant.enum";
 import { Organization } from "../organization/organization.model";
 import { Unit } from "../organization/organization.enum";
 import { User } from "../users/user.model";
+import { checkOrganizationOwnership } from "../../util/ownershipChecker";
 
 export interface GetApplicantsOptions {
     organization?: mongoose.Types.ObjectId;
@@ -40,7 +41,11 @@ export class ApplicantService {
         }
     }
 
-    static async createApplicant(data: CreateApplicantDto) {
+    static async createApplicant(data: CreateApplicantDto, userId: string) {
+        const ownsOrg = await checkOrganizationOwnership(userId, data.organization);
+        if (!ownsOrg) {
+            throw new Error("You are not authorized to create an applicant under this organization.");
+        }
         await this.validateApplicant(data);
         const createdApplicant = await Applicant.create({ ...data });
         return createdApplicant;
@@ -53,22 +58,30 @@ export class ApplicantService {
         return await Applicant.find(filter).populate('organization').lean();
     }
 
-    static async updateApplicant(id: string, data: Partial<CreateApplicantDto>) {
+    static async updateApplicant(id: string, data: Partial<CreateApplicantDto>, userId: string) {
         const applicant = await Applicant.findById(id);
         if (!applicant) throw new Error("Applicant not found");
+        const ownsOrg = await checkOrganizationOwnership(userId, applicant.organization);
+        if (!ownsOrg) {
+            throw new Error("You are not authorized to create an applicant under this organization.");
+        }
         Object.assign(applicant, data);
         return await applicant.save();
     }
 
-    static async deleteApplicant(id: string) {
+    static async deleteApplicant(id: string, userId: string) {
         const applicant = await Applicant.findById(id);
         if (!applicant) throw new Error("Applicant not found");
+        const ownsOrg = await checkOrganizationOwnership(userId, applicant.organization);
+        if (!ownsOrg) {
+            throw new Error("You are not authorized to create an applicant under this organization.");
+        }
         return await applicant.deleteOne();
     }
 
     static async autoLinkUserByEmail(applicantId: string) {
         const applicant = await Applicant.findById(applicantId);
-        if (!applicant) throw new Error("Applicant not found");
+        if (!applicant) throw new Error("Applicant not found");        
         if (!applicant.email) throw new Error("Applicant has no email to link");
         if (applicant.user) throw new Error("This user is already linked to another applicant");
         const user = await User.findOne({ email: applicant.email });
